chore: move NewTracer to keeper
